+++ src/teehr/loading/nwm22/nwm_point_data.py
@@ -26,9 +26,13 @@ def file_chunk_loop(
     variable_name: str,
     configuration: str,
     schema: pa.Schema,
+    ignore_missing_file: bool,
 ):
     """Fetch NWM values and convert to tabular format for a single json"""
-    ds = get_dataset(row.filepath).sel(feature_id=location_ids)
+    ds = get_dataset(row.filepath, ignore_missing_file)
+    if not ds:
+        return None
+    ds = ds.sel(feature_id=location_ids)
     vals = ds[variable_name].astype("float32").values
     nwm22_units = ds[variable_name].units
     teehr_units = NWM22_UNIT_LOOKUP.get(nwm22_units, nwm22_units)
@@ -62,7 +66,8 @@ def process_chunk_of_files(
     configuration: str,
     variable_name: str,
     output_parquet_dir: str,
-    process_by_z_hour: bool
+    process_by_z_hour: bool,
+    ignore_missing_file: bool,
 ) -> None:
     """Assemble a table for a chunk of NWM files"""
 
@@ -84,10 +89,21 @@ def process_chunk_of_files(
     for row in df.itertuples():
         results.append(
             file_chunk_loop(
-                row, location_ids, variable_name, configuration, schema
+                row,
+                location_ids,
+                variable_name,
+                configuration,
+                schema,
+                ignore_missing_file
             )
         )
     output = dask.compute(*results)
+
+    if not any(output):
+        raise FileNotFoundError("No NWM files for specified input"
+                                "configuration were found in GCS!")
+
+    output = [tbl for tbl in output if tbl is not None]
     output_table = pa.concat_tables(output)
 
     if process_by_z_hour:
@@ -114,6 +130,7 @@ def fetch_and_format_nwm_points(
     output_parquet_dir: str,
     process_by_z_hour: bool,
     stepsize: int,
+    ignore_missing_file: bool,
 ):
     """Reads in the single reference jsons, subsets the
         NWM data based on provided IDs and formats and saves
@@ -136,6 +153,11 @@ def fetch_and_format_nwm_points(
         for processing.
     stepsize: int
         The number of json files to process at one time.
+    ignore_missing_file: bool
+        Flag specifying whether or not to fail if a missing NWM
+        file is encountered
+        True = skip and continue
+        False = fail
     """
 
     output_parquet_dir = Path(output_parquet_dir)
@@ -152,7 +174,6 @@ def fetch_and_format_nwm_points(
     df_refs = pd.DataFrame(
         {"day": days, "z_hour": z_hours, "filepath": json_paths}
     )
-
     if process_by_z_hour:
         # Option #1. Groupby day and z_hour
         gps = df_refs.groupby(["day", "z_hour"])
@@ -173,6 +194,7 @@ def fetch_and_format_nwm_points(
             variable_name,
             output_parquet_dir,
             process_by_z_hour,
+            ignore_missing_file,
         )
 
 
@@ -186,8 +208,9 @@ def nwm_to_parquet(
     json_dir: str,
     output_parquet_dir: str,
     t_minus_hours: Optional[Iterable[int]] = None,
-    process_by_z_hour=True,
-    stepsize=100,
+    process_by_z_hour: Optional[bool] = True,
+    stepsize: Optional[int] = 100,
+    ignore_missing_file: Optional[bool] = True
 ):
     """Fetches NWM point data, formats to tabular, and saves to parquet
 
@@ -227,6 +250,10 @@ def nwm_to_parquet(
         The number of json files to process at one time. Used if
         process_by_z_hour is set to False. Default value is 100. Larger values
         can result in greater efficiency but require more memory
+    ignore_missing_file: bool
+        Flag specifying whether or not to fail if a missing NWM file is encountered
+        True = skip and continue
+        False = fail
 
     The NWM configuration variables, including configuration, output_type, and
     variable_name are stored as pydantic models in point_config_models.py
@@ -254,9 +281,12 @@ def nwm_to_parquet(
         start_date,
         ingest_days,
         t_minus_hours,
+        ignore_missing_file,
     )
 
-    json_paths = build_zarr_references(component_paths, json_dir)
+    json_paths = build_zarr_references(component_paths,
+                                       json_dir,
+                                       ignore_missing_file)
 
     fetch_and_format_nwm_points(
         json_paths,
@@ -266,17 +296,18 @@ def nwm_to_parquet(
         output_parquet_dir,
         process_by_z_hour,
         stepsize,
+        ignore_missing_file,
     )
 
 
 if __name__ == "__main__":
     configuration = (
-        "analysis_assim_extend"  # analysis_assim_extend, short_range
+        "short_range"  # analysis_assim_extend, short_range
     )
     output_type = "channel_rt"
     variable_name = "streamflow"
-    start_date = "2023-03-18"
-    ingest_days = 3
+    start_date = "2023-08-24"
+    ingest_days = 2
     location_ids = [
         7086109,
         7040481,
@@ -295,8 +326,9 @@ if __name__ == "__main__":
     json_dir = "/mnt/sf_shared/data/ciroh/jsons"
     output_parquet_dir = "/mnt/sf_shared/data/ciroh/parquet"
 
-    process_by_z_hour = False
+    process_by_z_hour = True
     stepsize = 100
+    ignore_missing_file = False
 
     nwm_to_parquet(
         configuration,
@@ -307,7 +339,8 @@ if __name__ == "__main__":
         location_ids,
         json_dir,
         output_parquet_dir,
-        t_minus_hours=[0],
+        t_minus_hours=[0, 1, 2],
         process_by_z_hour=process_by_z_hour,
         stepsize=stepsize,
+        ignore_missing_file=ignore_missing_file
     )