+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.4] - 2023-08-30
+
+### Changed
+
+- Behavior of loading when encountering missing files
+- Renamed field `zone` to `location_id` in `nwm_grid_data.py` and `generate_weights.py`
+
+### Added
+- The boolean flag `ignore_missing_files` to point and grid loading to determine whether to fail or continue on missing NWM files
+- Added a check to skip locally existing zarr json files when loading NWM data
+
 ## [0.2.3] - 2023-08-23
 
 ### Changed